+++ src/rer/newsletter/adapter/sender.py
@@ -6,7 +6,6 @@ from persistent.dict import PersistentDict
 from persistent.list import PersistentList
 from plone import api
 from rer.newsletter import logger
-from rer.newsletter.behaviors.ships import IShippable
 from rer.newsletter.utils import get_site_title
 from rer.newsletter.utils import OK
 from rer.newsletter.utils import SEND_UID_NOT_FOUND
@@ -73,27 +72,16 @@ class BaseAdapter(object):
 
     def _getMessage(self, message, footer):
         logger.debug("getMessage %s %s", self.context.title, message.title)
-        content = IShippable(message).message_content
-        message_template = self.context.restrictedTraverse("@@messagepreview_view")
+        message_template = message.restrictedTraverse("@@messagepreview_view")
         parameters = {
-            "css": self.context.css_style,
-            "message_header": self.context.header if self.context.header else "",
             "message_subheader": f"""
                 <tr>
                     <td align="left" colspan="2">
                       <div class="newsletterTitle">
-                        <h1>{self.context.title}</h1>
+                        <h1>{message.title}</h1>
                       </div>
                     </td>
                 </tr>""",
-            "message_footer": self.context.footer if self.context.footer else "",
-            "message_content": f"""
-                <tr>
-                    <td align="left" colspan="2">
-                    {content}
-                    </td>
-                </tr>
-            """,
             "message_unsubscribe_default": footer,
         }
 
@@ -136,14 +124,29 @@ class BaseAdapter(object):
         send_info = [x for x in details if x["uid"] == send_uid]
         if not send_info:
             return SEND_UID_NOT_FOUND
-        send_info[0]["send_date_end"] = datetime.today().strftime("%d/%m/%Y %H:%M:%S")
+        send_info[0]["send_date_end"] = datetime.today().strftime(
+            "%d/%m/%Y %H:%M:%S"
+        )
         send_info[0]["completed"] = completed
         send_info[0]["running"] = False
         return OK
 
     def prepare_body(self, message):
-        unsubscribe_footer_template = self.context.restrictedTraverse(
-            "@@unsubscribe_channel_template"
+        # unsubscribe_footer_template = self.context.restrictedTraverse(
+        #     "@@unsubscribe_channel_template"
+        # )
+        # parameters = {
+        #     "portal_name": get_site_title(),
+        #     "channel_name": self.context.title,
+        #     "unsubscribe_link": self.context.absolute_url(),
+        #     "enabled": self.context.standard_unsubscribe,
+        # }
+        # footer = unsubscribe_footer_template(**parameters)
+        # return self._getMessage(message=message, footer=footer)
+        unsubscribe_footer_template = api.content.get_view(
+            name="unsubscribe_channel_template",
+            context=self.context,
+            request=self.request,
         )
         parameters = {
             "portal_name": get_site_title(),
@@ -151,8 +154,36 @@ class BaseAdapter(object):
             "unsubscribe_link": self.context.absolute_url(),
             "enabled": self.context.standard_unsubscribe,
         }
-        footer = unsubscribe_footer_template(**parameters)
-        return self._getMessage(message=message, footer=footer)
+
+        message_template = api.content.get_view(
+            context=message, name="messagepreview_view", request=self.request
+        )
+        parameters = {
+            "message_subheader": f"""
+                <tr>
+                    <td align="left" colspan="2">
+                      <div class="newsletterTitle">
+                        <h1>{message.title}</h1>
+                      </div>
+                    </td>
+                </tr>""",
+            "message_unsubscribe_default": f"""
+                <tr>
+                    <td align="left" colspan="2">
+                    <div class="newsletter-unsubscribe">
+                        {unsubscribe_footer_template(**parameters)}
+                    </div>
+                    </td>
+                </tr>
+            """,
+        }
+
+        body = message_template(**parameters)
+        # passo la mail per il transform
+        portal = api.portal.get()
+        body = portal.portal_transforms.convertTo("text/mail", body)
+
+        return body
 
     def sendMessage(self, message):
         """This is the primary method to send emails for the channel."""
@@ -161,7 +192,9 @@ class BaseAdapter(object):
         subscribers = self.get_annotations_for_channel(key=SUBSCRIBERS_KEY)
 
         nl_subject = (
-            " - " + self.context.subject_email if self.context.subject_email else ""
+            " - " + self.context.subject_email
+            if self.context.subject_email
+            else ""
         )
         sender = (
             self.context.sender_name
@@ -207,7 +240,9 @@ class BaseAdapter(object):
     def addToHistory(self, message):
         """Add to history that message is sent"""
 
-        list_history = [x for x in message.workflow_history.get("message_workflow")]
+        list_history = [
+            x for x in message.workflow_history.get("message_workflow")
+        ]
         current = api.user.get_current()
         entry = dict(
             action="Invio",