+++ src/rer/newsletter/restapi/services/channel_subscriptions/unsubscribe.py
@@ -4,13 +4,15 @@ from plone.protect import interfaces
 from plone.protect.authenticator import createToken
 from plone.restapi.deserializer import json_body
 from plone.restapi.services import Service
-from rer.newsletter import logger
+from rer.newsletter import _
 from rer.newsletter.adapter.subscriptions import IChannelSubscriptions
 from rer.newsletter.utils import compose_sender
 from rer.newsletter.utils import get_site_title
+from rer.newsletter.utils import INEXISTENT_EMAIL
 from rer.newsletter.utils import OK
 from rer.newsletter.utils import UNHANDLED
 from six import PY2
+from zExceptions import BadRequest
 from zope.component import getMultiAdapter
 from zope.interface import alsoProvides
 
@@ -18,51 +20,54 @@ from zope.interface import alsoProvides
 class NewsletterUnsubscribe(Service):
     def reply(self):
         data = json_body(self.request)
+        email = data.get("email", "").lower()
+        if not email:
+            raise BadRequest(
+                api.portal.translate(
+                    _(
+                        "missing_email_label",
+                        default="Missing required parameter: email.",
+                    )
+                )
+            )
+
         if "IDisableCSRFProtection" in dir(interfaces):
             alsoProvides(self.request, interfaces.IDisableCSRFProtection)
-        response, errors = self.handleUnsubscribe(data)
-        if errors:
-            response["errors"] = errors
-        return response
 
-    def getData(self, data):
-        errors = []
-        if not data.get("email", None):
-            errors.append("Indirizzo email non inserito o non valido")
-        return {
-            "email": data.get("email", None),
-        }, errors
+        self.handle_unsubscribe(email=email)
 
-    def handleUnsubscribe(self, postData):
+        return self.reply_no_content()
+
+    def handle_unsubscribe(self, email):
         status = UNHANDLED
-        data, errors = self.getData(postData)
-        if errors:
-            return data, errors
-
-        email = data.get("email", None)
 
         channel = getMultiAdapter((self.context, self.request), IChannelSubscriptions)
 
         status, secret = channel.unsubscribe(email)
-
         if status != OK:
-            logger.exception("Error: {}".format(status))
-            if status == 4:
-                msg = "unsubscribe_inexistent_mail"
-
-            else:
-                msg = "unsubscribe_generic"
-            errors = msg
-            return {"@id": self.request.get("URL")}, errors
+            if status == INEXISTENT_EMAIL:
+                raise BadRequest(
+                    api.portal.translate(
+                        _(
+                            "unsubscribe_inexistent_mail",
+                            default="Mail not found. Unable to unsubscribe.",
+                        )
+                    )
+                )
+            raise BadRequest(
+                api.portal.translate(
+                    _(
+                        "unsubscribe_generic",
+                        default="Unable to perform unsubscription. Please contact site administrators.",
+                    )
+                )
+            )
 
         # creo il token CSRF
         token = createToken()
 
         # mando mail di conferma
-        url = self.context.absolute_url()
-        url += "/confirm-subscription?secret=" + secret
-        url += "&_authenticator=" + token
-        url += "&action=unsubscribe"
+        url = f"{self.context.absolute_url()}/confirm-subscription?secret={secret}&_authenticator={token}&action=unsubscribe"
 
         mail_template = self.context.restrictedTraverse("@@deleteuser_template")
 
@@ -89,17 +94,8 @@ class NewsletterUnsubscribe(Service):
             mail_text.getData(),
             mto=email,
             mfrom=response_email,
-            subject="Conferma la cancellazione dalla newsletter"
-            " {channel} del portale {site}".format(
-                channel=channel_title, site=get_site_title()
-            ),
+            subject=f"Conferma la cancellazione dalla newsletter {channel_title} del portale {get_site_title()}",
             charset="utf-8",
             msg_type="text/html",
             immediate=True,
         )
-
-        return {
-            "@id": self.request.get("URL"),
-            "status": "user_unsubscribe_success" if not errors else "error",
-            "errors": errors if errors else None,
-        }, errors